services/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
pytest-asyncio==0.21.1